refactor(js): 模块化 js 代码
