+++ src/main/resources/static/js/modules/siteCharts/constants.js
@@ -1,0 +1,40 @@
+export const API_ENDPOINT = '/apis/api.data.statistics.xhhao.com/v1alpha1/chart/data';
+
+export const COLOR_PALETTE = [
+    '#3b82f6', '#10b981', '#f59e0b', '#ef4444', '#8b5cf6',
+    '#ec4899', '#14b8a6', '#f97316', '#6366f1', '#0ea5e9'
+];
+
+export const BAR_SOFT_COLORS = [
+    'rgba(255, 99, 132, 0.22)',
+    'rgba(255, 159, 64, 0.22)',
+    'rgba(255, 205, 86, 0.22)',
+    'rgba(75, 192, 192, 0.22)',
+    'rgba(54, 162, 235, 0.22)',
+    'rgba(153, 102, 255, 0.22)',
+    'rgba(201, 203, 207, 0.22)',
+    'rgba(236, 72, 153, 0.22)',
+    'rgba(16, 185, 129, 0.22)',
+    'rgba(14, 165, 233, 0.22)'
+];
+
+export const BAR_STRONG_COLORS = [
+    'rgb(255, 99, 132)',
+    'rgb(255, 159, 64)',
+    'rgb(255, 205, 86)',
+    'rgb(75, 192, 192)',
+    'rgb(54, 162, 235)',
+    'rgb(153, 102, 255)',
+    'rgb(201, 203, 207)',
+    'rgb(236, 72, 153)',
+    'rgb(16, 185, 129)',
+    'rgb(14, 165, 233)'
+];
+
+export const DAY_IN_MS = 24 * 60 * 60 * 1000;
+
+export const WEEKDAY_LABELS = ['一', '二', '三', '四', '五', '六', '日'];
+
+export const MONTH_NAMES = ['一月', '二月', '三月', '四月', '五月', '六月', '七月', '八月', '九月', '十月', '十一月', '十二月'];
+
+export const DEFAULT_CHART_TYPES = ['tags', 'categories', 'articles', 'comments', 'topArticles'];
